requirements.txt
onnxruntime
streamlit
numpy
pillow
streamlit-drawable-canvas